Labels wider than a byte, edge cells removed before the batch cell alignment, and the Z bound the runs actually use
Three things found while putting two synchronised RNA experiments
(JP_001, JP_002: 30 FOVs each, brightfield segmentation) through the
whole pipeline headless.

THE LABEL WRAP. segment._filter_and_relabel returned the relabelled
mask as uint8. A dense brightfield field holds more than 255 cells:
MEASURED on JP_001, 12 of 30 FOVs reported exactly 255 cells, because
label 256 became background and 257 onward wrapped onto cells 1, 2,
... -- a distant cell's pixels silently became part of another cell's
mask. Now uint16, like _merge_append_mask already was; every consumer
builds per-cell boolean masks from the labels and never depended on the
width. tests/test_segment_relabel.py keeps 300 cells apart. Re-segmented
JP_001 went from 6409 to 7030 cells.

EDGE CELLS, AS A BATCH. CellContainer.edge_cell_ids applies the Cell
Displayer's Remove Edge Cells criterion (a label on the first or last
row or column) to the container's own native masks, so a batch can use
it without a rendered mask. "Align All Cells in ALL FOVs" gains a
checkbox, on by default and stored as cell_alignment.remove_edge_cells:
before fitting, every edge cell is removed from every FOV in the list,
the other tier is kept in step, the FOV is written to every store and
its spots recast -- the three steps a manual removal followed by Save
performs. A FOV emptied by the removal drops out of the run. On JP_001
that removed 1058 of 7030 cells; on JP_002, 985 of 6636.
tests/test_edge_cells.py.

THE Z BOUND. chain.MAX_CELL_Z_SHIFT_PLANES was 15 while every real
config carried 55 and a 15-plane cap had truncated real shifts on the
first project it met; the panel's default now says what the runs use.
The two JP runs fitted with 55: |dz| p90 6 planes, max 42.

Also the two configs for those experiments (RNA only, Hyb_101
reference, Hyb_500 brightfield Cellpose, FOV-mode celltypes
Unsynchronized / Hydroxyurea / Nocodazole).

+++ codelab_pipeline/segmentation/segment.py
@@ -153,7 +153,15 @@ def _filter_and_relabel(mask, min_size, max_size):
     mask[np.isin(mask, v[c > max_size])] = 0
     for i, id in enumerate(np.unique(mask)[1:], start=1):
         mask[mask == id] = -i
-    return (-1 * mask).astype(np.uint8)
+    # uint16, NOT uint8. The label image is the cell identity, and a
+    # dense field holds more than 255 cells: MEASURED on JP_001 (BF,
+    # Cellpose diameter 40, 1024x1024), 12 of 30 FOVs reported exactly
+    # 255 cells because label 256 wrapped to background and 257 onward
+    # wrapped onto cells 1, 2, ... -- a distant cell's pixels silently
+    # became part of another cell's mask. _merge_append_mask already
+    # returns uint16 for the same reason; every consumer builds per-cell
+    # boolean masks from the labels and never depends on the width.
+    return (-1 * mask).astype(np.uint16)
 
 
 def segment_fov_classical(storage_path, fov, reference_hybe, channel, method='otsu',
